moves subgroup components to common to avoid dependency on lazy module
