+++ webapp/src/main/webapp/src/app/aggregate-report/subgroup/subgroup-filter-options.ts
@@ -1,3 +1,3 @@
-import { SubgroupFilters } from './subgroup-filters';
+import { SubgroupFilters } from '../../shared/model/subgroup-filters';
 
 export interface SubgroupFilterOptions extends SubgroupFilters {}
