--- webapp/src/main/webapp/src/app/aggregate-report/subgroup/subgroup-item.ts
@@ -1,7 +1,0 @@
-import { SubgroupFilters } from './subgroup-filters';
-import { Subgroup } from './subgroup';
-
-export interface SubgroupItem {
-  readonly source: SubgroupFilters;
-  readonly subgroup: Subgroup;
-}
