+++ webapp/src/main/webapp/src/app/aggregate-report/subgroup/subgroup.component.ts
@@ -1,5 +1,5 @@
 import { Component, Input } from '@angular/core';
-import { Subgroup } from './subgroup';
+import { Subgroup } from '../../shared/model/subgroup';
 
 @Component({
   selector: 'subgroup',
