+++ webapp/src/main/webapp/src/app/shared/model/subgroup-item.ts
@@ -1,0 +1,7 @@
+import { SubgroupFilters } from './subgroup-filters';
+import { Subgroup } from './subgroup';
+
+export interface SubgroupItem {
+  readonly source: SubgroupFilters;
+  readonly subgroup: Subgroup;
+}
